move logic from main to path
